Seperated GUI and server in the server application
Seperated the GUI and the server code in the server application.

The RegistryServer is a separate class now. The RegistryServer class has
an output field that logs all the messages of the server. This output
field is observable, allowing the connection of multiple GUIs with
different outputting mechanisms to display the output properly by
observing the output field.