HostedGamesController JAT offloaded
Seperated the Clash of Fishes server lookup and GUI changes from the
initialize and put it in seperate methods.

Simplified the calls to the server lookup and lobby list refresh by
adding the 'getNewLobbies' method that creates a new thread to offload
the JAT from having to communicate with an external network IO entity.